requirements.txt
django
django-extensions==3.2.3
hl7
django-phone-field
